would be cool to allow automatic cleaning after Ctrl-C (aka make net global)

+++ mn_test.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env nix-shell 
 #!nix-shell shell-mininet.nix -i python --show-trace
 
+# To clean everything run 'mn -c'
+
 # python needs to read this 
 # -*- coding: utf-8 -*-
 # but it will check just the first 2 lines :/ https://www.python.org/dev/peps/pep-0263/#defining-the-encoding
@@ -62,10 +64,15 @@ topo = [
 #               params1=par1,
 #               params2=par2)
 
+net = None
+
 # clean sthg
 def sigint_handler(signum, frame):
     print('Stop pressing the CTRL+C!')
- 
+    global net
+    net.stop()
+    sys.exit(3)
+
 signal.signal(signal.SIGINT, sigint_handler)
 
 class StaticTopo(Topo):
@@ -89,6 +96,12 @@ class StaticTopo(Topo):
 # };
 
 def runExperiment(number_of_paths, interactive, test, loss, out, **kwargs):
+
+    def _out(*args):
+        """ Use it to name files"""
+        suffix = '_'.join(map(str, args))
+        return os.path.join(out, suffix )
+
     # using 
     net = Mininet(topo=StaticTopo(number_of_paths, loss), link=AsymTCLink)
     net.start()
@@ -104,8 +117,9 @@ def runExperiment(number_of_paths, interactive, test, loss, out, **kwargs):
     
     # heat network to avoid intial packet artefacts
     # for now I don't care
-    for i in range(number_of_paths):
-        client.cmd("ping 1" + str(i) + ".0.0.2 -c 4")
+    # for i in range(number_of_paths):
+    #     client.cmd("ping 1" + str(i) + ".0.0.2 -c 4")
+
         # client.cmd("owping 1" + str(i) + ".0.0.2 -c 4 2>&1 > owping" + str(i) + ".log ")
 
     # sys.exit(1)
@@ -128,8 +142,9 @@ def runExperiment(number_of_paths, interactive, test, loss, out, **kwargs):
     if kwargs.get("capture"):
         print("Capturing packets...")
         print('tshark -r out/client_2.pcap -z "conv,mptcp"')
-        client.cmd('tshark -i any -w out/client_' + str(number_of_paths) + '.pcap &')
-        server.cmd('tshark -i any -w out/server_' + str(number_of_paths) + '.pcap &')
+        # TODO use popen instead ?
+        client.cmd("tshark -i any -w '%s' &" % _out("client_", number_of_paths, ".pcap"))
+        server.cmd("tshark -i any -w '%s' &" % _out("server_", number_of_paths, ".pcap"))
         # let tshark the time to setup itself
         os.system("sleep 5")
     
@@ -149,29 +164,62 @@ def runExperiment(number_of_paths, interactive, test, loss, out, **kwargs):
     # client.cmd('flent rrul -p ping_cdf -l 60 -H 10.0.0.2 -t "mon titre" -o filename.png')   
 
     # # iperf 3 version
-    server_iperf = server.popen("iperf3 -s --json --logfile 'out/server_%d.log' &" % number_of_paths)
-    client.cmd("iperf -c {serverIP} -n {dataAmount} -i 1 > out/client_{paths}.log".format(
-        # server.IP() should work ok
-        serverIP="10.0.0.2",
-        dataAmount="5M",
-        paths=number_of_paths
-    ))
+    cmd = "iperf3 -s --json --logfile=%s" % (_out("server_iperf", number_of_paths, ".log"),)
+    print("starting", cmd)
+    server_iperf = server.popen(cmd)
+    # out, err = server_iperf.communicate()
+    # if err is not None:
+    if server_iperf.returncode:
+        print("Failed to run ", cmd)
+        print("returned", server_iperf.returncode)
+        print(err)
+        sys.exit(1)
+
+    # server.cmd("iperf3 -s --json --logfile '%s' &" % _out("server_iperf", number_of_paths, ".log"))
+    # server_iperf.poll()
     # TODO get results else it might get dirty
         
     # run_tests()
-        # TODO move the loop to here
+    # TODO move the loop to here
     for run in range(4):
 
-        def _out(suffix):
-            return os.path.join(out, suffix)
+        # in iperf3, the client sends the data so...
+        reinject_out = _out("check", run, ".csv")
+        print(reinject_out)
+        with open(reinject_out, "w+") as fd:
+            client_check = client.popen(
+                ["/home/teto/testbed/check_opportunistic_reinject.py", "-j"], 
+                stdout=fd,
+                universal_newlines=True
+            )
+            out, err = client_check.communicate()
 
+            if err is not None:
+                print("Failed running with returncode=", client_check.returncode)
+                print(err)
+                break
+            # assert err == 0
+
+            client_iperf = client.popen("iperf3 -c {serverIP} -n {dataAmount} --logfile {logfile} ".format(
+            # client.cmd("iperf3 -c {serverIP} -n {dataAmount} --logfile {logfile} ".format(
+                serverIP=server.IP(),   # seems to work
+                # serverIP="10.0.0.2",
+                dataAmount="5M",
+                # paths=number_of_paths
+                logfile=_out("client_iperf", "path", number_of_paths, "run", run, ".log")
+            ))
+            # wait for iperf to finish
+            client_iperf.wait()
+            # client.waitOutput()
+
+            # if client_check.returncode:
+            print("out/errs=", out, err)
+            print("returncode=", client_iperf.returncode)
+            # client.cmd("kill %d" % (client_check.pid,))
+            client_check.kill()
         
-        server.popen("/home/teto/testbed/check_opportunistic_reinject.py -j > %s", _out("check_%d.csv" % (run,)) )
-        client.waitOutput()
-        # server.kill("kill %while")
-
-        res = client.cmd('/home/teto/testbed/gen_cpt.sh 10.0.0.2')
-        print("RES=%r" % res)
+        # res = client.cmd('/home/teto/testbed/gen_cpt.sh 10.0.0.2')
+        # print("RES=%r" % res)
 
 
     if interactive:
@@ -181,18 +229,24 @@ def runExperiment(number_of_paths, interactive, test, loss, out, **kwargs):
     # lets wait a moment
     sleep(3)
     # and ensure iperf is finished :-)
-    server.cmd("kill %d" % (server_iperf.pid,))
+    # server.cmd("kill %d" % (server_iperf.pid,))
+    server_iperf.terminate()
     # os.system('pkill -f \'iperf\'')
-    os.system('pkill -f \'tshark\'')
+
+    if kwargs.get("capture"):
+        os.system('pkill -f \'tshark\'')
+
     net.stop()
     sleep(1)
   
 if __name__ == '__main__':
+
+    print("To clean run `mn -c`")
     parser = argparse.ArgumentParser()
     # parser.add_argument("-f", "--file", help="The file which contains the scheduler", required=True)
     parser.add_argument("-n", "--number_of_subflows", help="The number of subflows")
     parser.add_argument("-d", "--debug", choices=['debug', 'info', 'error'], help="Running in debug mode", default='info')
-    parser.add_argument("-t", "--test", choices=["iperf3-cdf"], help="capture packets", default="iperf3-cdf")
+    parser.add_argument("-t", "--test", choices=["iperf3-cdf"], help="test to run", default="iperf3-cdf")
     parser.add_argument("-c", "--capture", action="store_true", help="capture packets", default=False)
     parser.add_argument("-i", "--interactive", action="store_true", help="Waiting in command line interface", default=False)
     parser.add_argument("-l", "--loss", help="Loss rate (between 0 and 100", default=0)
@@ -203,6 +257,8 @@ if __name__ == '__main__':
     
     setLogLevel(args.debug)
 
+    print("CWD=", os.getcwd())
+    print("creating %s" % args.out)
     os.system("mkdir -p %s" % args.out)
     
     # if args.debug:
